feat(refactor): Furthers previous refactor, by moving
Functions to new module. Tidies up reused code into
private function.

Throws warning due to line 11 in command.rs - unused import
Import is used for the StructOpt derive macro, so is needed
but still throws this warning
