+++ tests/manifest/main.rs
@@ -28,7 +28,7 @@ fn it_gets_the_crate_name_provided_path() {
 #[test]
 fn it_creates_a_package_json_default_path() {
     let path = ".".to_string();
-    wasm_pack::create_pkg_dir(&path).unwrap();
+    wasm_pack::command::create_pkg_dir(&path).unwrap();
     assert!(manifest::write_package_json(&path, None).is_ok());
     let package_json_path = format!("{}/pkg/package.json", &path);
     assert!(fs::metadata(package_json_path).is_ok());
@@ -46,7 +46,7 @@ fn it_creates_a_package_json_default_path() {
 #[test]
 fn it_creates_a_package_json_provided_path() {
     let path = "tests/fixtures/js-hello-world".to_string();
-    wasm_pack::create_pkg_dir(&path).unwrap();
+    wasm_pack::command::create_pkg_dir(&path).unwrap();
     assert!(manifest::write_package_json(&path, None).is_ok());
     let package_json_path = format!("{}/pkg/package.json", &path);
     assert!(fs::metadata(package_json_path).is_ok());
@@ -59,7 +59,7 @@ fn it_creates_a_package_json_provided_path() {
 #[test]
 fn it_creates_a_package_json_provided_path_with_scope() {
     let path = "tests/fixtures/scopes".to_string();
-    wasm_pack::create_pkg_dir(&path).unwrap();
+    wasm_pack::command::create_pkg_dir(&path).unwrap();
     assert!(manifest::write_package_json(&path, Some("test".to_string())).is_ok());
     let package_json_path = format!("{}/pkg/package.json", &path);
     assert!(fs::metadata(package_json_path).is_ok());
